refactor: CreateTaskFormState to tasks/common
